refactor: extract localization into extension system
- Removed Localizable modules from Page, Site, and Finder
- Introduced Hotpages::Extensions::Localization for i18n logic
- Registered extension in Hotpages loader
- Simplified core classes by moving localization concerns to extension
